requirements.txt
python-telegram-bot[job-queue]
openai
groq
nltk
python-dotenv
